requirements.txt
flask
python-dotenv
werkzeug
groq
langchain-groq
langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf2
markdown2